refactor setup and kbar
