+++ src/components/Kbar/CommandBar/styles.module.css
@@ -1,0 +1,141 @@
+/* positioner */
+.positioner {
+  position: fixed;
+  display: flex;
+  align-items: flex-start;
+  justify-content: center;
+  width: 100%;
+  inset: 0px;
+  padding: 14vh 16px 16px;
+  background: var(--cb-bg-color);
+  box-sizing: border-box;
+  z-index: 4;
+}
+
+/* search */
+.search {
+  padding: 12px 16px;
+  font-size: 16px;
+  width: 100%;
+  box-sizing: border-box;
+  outline: none;
+  border: none;
+  margin: 0;
+  font-family: var(--font-secondary);
+  background: var(--sub-alt-color);
+  color: var(--text-color);
+
+
+}
+
+@supports ((-webkit-backdrop-filter: none) or (backdrop-filter: none)) {
+  .search {
+    background-color: #ffffff0d;
+    -webkit-backdrop-filter: saturate(207%) blur(30px);
+    backdrop-filter: saturate(207%) blur(30px);
+  }
+}
+
+/* groupname */
+.groupname {
+  padding: 8px 16px;
+  font-size: 10px;
+  text-transform: uppercase;
+  letter-spacing: 1px;
+  background: var(--sub-alt-color);
+  font-family: var(--font-secondary);
+
+
+}
+
+@supports ((-webkit-backdrop-filter: none) or (backdrop-filter: none)) {
+  .groupname {
+    background-color: #ffffff0d;
+    -webkit-backdrop-filter: saturate(207%) blur(30px);
+    backdrop-filter: saturate(207%) blur(30px);
+  }
+}
+
+/* kbd */
+.kbd {
+  background: var(--bg-alt-color);
+  color: var(--text-color);
+  padding: 4px 8px;
+  text-transform: uppercase;
+}
+
+/* shortcut */
+.shortcut {
+  display: grid;
+  grid-auto-flow: column;
+  gap: 4px;
+}
+
+/* action */
+.action {
+  display: flex;
+  gap: 8px;
+  align-items: center;
+  font-family: var(--font-secondary);
+  --fluid-type-min: .8rem;
+  --fluid-type-max: .9rem;
+  --fluid-type-target: 2vw;
+}
+
+.action svg {
+  color: var(--text-color);
+}
+
+/* actionrow */
+.actionrow {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+}
+
+.actionrow>span:first-child {
+  padding-right: 8px;
+}
+
+/* animator */
+.animator {
+  background-color: #ffffff0d;
+  max-width: 600px;
+  width: 100%;
+  color: var(--text-color);
+  border-radius: 8px;
+  overflow: hidden;
+  box-shadow: 0px 10px 20px 8px rgba(0, 0, 0, 0.2);
+
+
+}
+
+@supports ((-webkit-backdrop-filter: none) or (backdrop-filter: none)) {
+  .animator {
+    background-color: #ffffff0d;
+    -webkit-backdrop-filter: saturate(207%) blur(30px);
+    backdrop-filter: saturate(207%) blur(30px);
+  }
+}
+
+.animator>div>div::-webkit-scrollbar {
+  width: 3px;
+}
+
+.animator>div>div {
+  -ms-overflow-style: none;
+  scrollbar-width: none;
+}
+
+/* box */
+.box {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 8px 16px;
+  cursor: pointer;
+}
+
+.box:hover {
+  background-color: var(--cb-box-alt-color);
+}
