+++ src/domains/projects/YearProjects/index.tsx
@@ -1,7 +1,9 @@
+'use client'
+
+import { motion } from 'framer-motion'
 import { useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { ProjectItem, projects } from './data'
-import { motion } from 'framer-motion'
 import styles from './styles.module.css'
 
 export const YearProjects = () => {
