--- src/domains/setup/components/setupmapper/index.ts
@@ -1,1 +1,0 @@
-export { SetupMapper } from './setupmapper'
