+++ src/domains/setup/components/setupmapper/styles.module.css
@@ -1,0 +1,47 @@
+/* setupMapperContainer */
+.setupMapperContainer {
+  padding-top: 2rem;
+}
+
+.setupMapperContainer h2 {
+  color: var(--text-color);
+  font-family: var(--font-primary);
+  font-weight: 600;
+  font-size: 1.5rem;
+}
+
+.setupMapperContainer ul {
+  padding-left: 1.8rem;
+  padding-top: 1rem;
+}
+
+.setupMapperContainer ul li {
+  color: var(--text-color);
+  font-family: var(--font-primary);
+  line-height: 2;
+
+  --fluid-type-max: 1em;
+}
+
+.setupMapperContainer ul li span {
+  color: var(--text-color);
+  font-weight: 300;
+  font-family: var(--font-primary);
+}
+
+/* softwareLink */
+.softwareLink {
+  color: var(--text-color);
+  font-family: var(--font-primary);
+  line-height: 1.8;
+  font-weight: 500;
+  transition: var(--transition-primary);
+}
+
+.softwareLink:hover {
+  color: var(--main-color);
+}
+
+.softwareLink span {
+  color: var(--main-color);
+}
