--- src/domains/setup/components/setupmapper/styles.ts
@@ -1,44 +1,0 @@
-import { styled } from 'src/ui/theme'
-
-export const SetupMapperContainer = styled('section', {
-  paddingTop: '2rem',
-  h2: {
-    color: '$text_color',
-    fontFamily: 'var(--font-secondary)',
-    fontWeight: 600,
-    fontSize: '1.5rem',
-  },
-  ul: {
-    paddingLeft: '1.8rem',
-    paddingTop: '1rem',
-    li: {
-      color: '$text_color',
-      fontFamily: 'var(--font-secondary)',
-      lineHeight: '2',
-      '--fluid-type-min': '0.8rem',
-      '--fluid-type-max': '1rem',
-      span: {
-        color: '$text_color',
-        fontWeight: 300,
-      },
-    },
-  },
-})
-
-export const SoftwareLink = styled('a', {
-  color: '$text_color',
-  fontFamily: 'var(--font-secondary)',
-  lineHeight: '1.8',
-  fontWeight: 500,
-  transition: '$transitonTheme',
-  '--fluid-type-min': '0.8rem',
-  '--fluid-type-max': '1rem',
-
-  '&:hover': {
-    color: '$main_color',
-  },
-
-  span: {
-    color: '$main_color',
-  },
-})
