--- src/domains/setup/components/terminal/index.ts
@@ -1,1 +1,0 @@
-export { Terminal } from './terminal'
