--- src/domains/setup/features/search/index.ts
@@ -1,1 +1,0 @@
-export { Search } from './Search'
