--- src/domains/setup/styles.ts
@@ -1,15 +1,0 @@
-import { styled } from 'src/ui/theme'
-
-export const Main = styled('main', {
-  display: 'flex',
-  flexDirection: 'column',
-  justifyContent: 'center',
-  alignItems: 'flex-start',
-  transition: '$transitonTheme',
-  padding: '2rem 0',
-})
-
-export const Content = styled('div', {
-  paddingTop: '2rem',
-  width: '100%',
-})
